+++ waspc/examples/todoApp/src/rpcTests/operations/jsDefinitions.js
